requirements.txt
Flask
requests
python-dotenv
openai
numpy
pandas
tiktoken
langchain
faiss-cpu